// models/RevokeTokenRequest.ts
/**
 * Stellar DataStore
 * The Stellar DataStore service is your instantly available, secure and worry-free cloud data storage service. Sign-up now and minutes later, you can start putting your data in the cloud.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class RevokeTokenRequest {
    'refreshToken': string;
    'clientId': string;
    'clientSecret': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "refreshToken",
            "baseName": "refresh_token",
            "type": "string",
            "format": ""
        },
        {
            "name": "clientId",
            "baseName": "client_id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "clientSecret",
            "baseName": "client_secret",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RevokeTokenRequest.attributeTypeMap;
    }

    public constructor() {
    }
}