// models/UpdateRecordRequest.ts
/**
 * Stellar DataStore
 * The Stellar DataStore service is your instantly available, secure and worry-free cloud data storage service. Sign-up now and minutes later, you can start putting your data in the cloud.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class UpdateRecordRequest {
    'idList'?: Array<any> | null;
    'record': { [key: string]: any | null; };

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "idList",
            "baseName": "idList",
            "type": "Array<any>",
            "format": ""
        },
        {
            "name": "record",
            "baseName": "record",
            "type": "{ [key: string]: any | null; }",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return UpdateRecordRequest.attributeTypeMap;
    }

    public constructor() {
    }
}